requirements.txt
gradio>=4.44.0
python-dotenv==1.0.1

spacy==3.7.6
nltk==3.9.1
langdetect==1.0.9

pypdf==5.1.0
python-docx==1.1.2

rapidfuzz==3.10.1
pydantic==2.10.3
requests==2.32.3
